+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: validate uploaded SSH key pairs;
 - concord-server, concord-agent: support for pulling dependencies
 from Maven repositories;
 - concord-server: update repositories using GitHub webhooks;